Completed all the question leaving one (Substring in a String), Added Solutions to StringComparison, StringConcatenation, StringPalindrome, StringReverse, StringToLower, StringToUpper. Modified th ReverseString to add another way to do the same, basically using Swapping.
Now, Done the String Comparison in both using library/function and Manually by checking and comparing each charecter in the array,
Done the Concatenation in ways using a third array or directly into the first array,
Checked for Palindrome using reversing the String manually using Swapping method,
Reverse is done using the old way by Swapping the Original String using indexes,
Finally, toLower is being done by converting a Charecter to ASCII then checking if it exists in the range of (65 - 90) of Upper Values if yes, then get the difference and Add that to 65 to get the Same Alphabet position in Lower ASCII, finally convert the lowerASCII to corresponding Charecter and Swap it in the Original String, Same for the toUpper.
Only the SubString in a String Question Solution is left.